Added rest of neccessary files for the assignment (currently blank). Added information to the HPP file Adjusted for boolean values Commetns added to ensure readability of Box.hpp Typed up and readded information for the boolean values Added the main constructor value for when the user adds in values at creation of the object

+++ Box.hpp
@@ -16,9 +16,9 @@
 class Box {
 
 private:
-    int height;
-    int width;
-    int length;
+    double height;
+    double width;
+    double length;
 
 public:
     // Constructor in case the user doesn't put anything in.
